Cambio de Integer a double para el precio de los productos, de forma que ahora se mostrarán los céntimos.

+++ src/ESListaDeProductos.java
@@ -1,6 +1,6 @@
 public class ESListaDeProductos {
     public static void insertaProducto(String nombre, double precio, ListaDeProductos lista){
-        lista.insertaProducto(nombre, (int)precio);
+        lista.insertaProducto(nombre, precio);
         lista.actualizaLista();
     }
 }
